Create System.Net.ServicePoint assembly
- Creates a new System.Net.ServicePoint contract
- Moves the relevant typs down into it from System.Net.Requests
- Deletes the obsolete ICertificatePolicy that was preventing this separation
