--- src/System.Net.Requests/src/System/Net/ICertificatePolicy.cs
@@ -1,13 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Security.Cryptography.X509Certificates;
-
-namespace System.Net
-{
-    public interface ICertificatePolicy
-    {
-        bool CheckValidationResult(ServicePoint srvPoint, X509Certificate certificate, WebRequest request, int certificateProblem);
-    }
-}
