+++ src/System.Net.ServicePoint/ref/project.json
@@ -1,0 +1,11 @@
+{
+  "dependencies": {
+    "System.Net.Primitives": "4.4.0-beta-24619-02",
+    "System.Net.Security": "4.4.0-beta-24619-02",
+    "System.Runtime": "4.4.0-beta-24619-02",
+    "System.Security.Cryptography.X509Certificates": "4.4.0-beta-24619-02"
+  },
+  "frameworks": {
+    "netstandard1.7": {}
+  }
+}
