+++ src/System.Net.ServicePoint/src/System/Net/SecurityProtocolType.cs
