+++ src/System.Net.ServicePoint/src/project.json
@@ -1,0 +1,23 @@
+{
+  "frameworks": {
+    "netstandard1.7": {
+      "dependencies": {
+        "Microsoft.NETCore.Platforms": "1.0.1",
+        "System.Collections": "4.4.0-beta-24619-02",
+        "System.Collections.Concurrent": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Debug": "4.4.0-beta-24619-02",
+        "System.Net.Primitives": "4.4.0-beta-24619-02",
+        "System.Net.Security": "4.4.0-beta-24619-02",
+        "System.Resources.ResourceManager": "4.4.0-beta-24619-02",
+        "System.Runtime": "4.4.0-beta-24619-02",
+        "System.Runtime.Extensions": "4.4.0-beta-24619-02",
+        "System.Security.Cryptography.X509Certificates": "4.4.0-beta-24619-02"
+      }
+    },
+    "net463": {
+      "dependencies": {
+        "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
+      }
+    }
+  }
+}
