+++ src/System.Net.WebClient/tests/System.Net.WebClient.Tests.csproj
@@ -33,6 +33,7 @@
     <!-- TODO: Remove these reference once packages are updated -->
     <ProjectReference Include="..\..\System.Net.Primitives\pkg\System.Net.Primitives.pkgproj" />
     <ProjectReference Include="..\..\System.Net.Requests\pkg\System.Net.Requests.pkgproj" />
+    <ProjectReference Include="..\..\System.Net.ServicePoint\pkg\System.Net.ServicePoint.pkgproj" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
